remove unnecessary file
